Fuzzing scriptSig working nicely

+++ src/fuzzer.h
@@ -6,7 +6,7 @@
 
 #include "main.h"
 
-extern void FuzzTransaction(const CTransaction& tx, const uint32_t& fuzzSeed, CDataStream& fuzzedDataRet);
+extern void FuzzTransaction(const CTransaction& tx, const uint64_t& fuzzSeed, CDataStream& fuzzedDataRet);
 extern void FuzzRelayTransaction(const CTransaction& tx);
 
 #endif
